Made the docker hostable.
Created an nginx container that builds the static html and sets up proxying to the backend. nginx provides authentication and ssl

Modified the python files to use an ssh key stored as an environment variable

Modified the frontend to use local paths. Removed the requirement for typescript to compile to build.